Remove undefined null initializations from GET_ITEM macros
These macros used to initialize GetItemEntry types with NULL for drawItemId.
While it's often allowed to initialize integer literals with NULL (as
opposed to nullptr), it's frowned upon and disallowed by the core
guidelines. Compilers are allowed to warn on this conversion and depending
on the implementation may even fail (e.g. if it's typedef'd to nullptr).
The integer literal is more appropriate to use.

Additionally, these values should properly be set to a copy of itemId and
modIndex in most cases as per comments.