Choose reasonable defaults for the GamePlatform and GameRegion
This is mainly to ensure that the relevant functions don't go into
undefined behavior land if something is awry. It'd be better to have
getGameVersions() return an enum the compiler can ensure is completely
handled or we should throw an error, but frankly I'm not sure how to do
either of those from this context.

Good enough for now.