Another signed comparison issue.
I'd like to rewrite this function to use an owning type for the duplicated
strings, but doing that while keeping the constraints of extern "C" is
tricky. Best would probably be a std::vector<std::vector>> or the equivalent
with std::array, but the lifetimes are dangerous once we start passing
mutable references to C code. For now, we fix the immediate signed
comparison issue and leave this little mess be, comments and all.