Extremely sketchy hack
Most of these defines are defined before this point by system headers.
We really shouldn't be defining them at all (it'd be better to check if
they're already defined and not overwrite them if so). However, I'm
not sure if these are the values expected to match existing systems so
instead I'm undef'ing the existing values. Yikes.