Fix an ordering issue an UB with MessageNotFoundException
When MessageNotFoundException is printed, the string return is written
into an uninitialized pointer on the stack. Instant stack corruption.

This commit fixes the issue by constructing the format string at init.
This allows us to remove the additional data that was being stored. To
avoid redundancy, I removed the move constructor because it's probably
unnecessary here.