Work around a difficult compatibility issue between C/C++ with enums
The ItemID enum type in z64Item.h is guaranteed to be sufficient to hold
the enum values specified. However, the s8 used by the debug save editor
can't represent the value of ITEM_NONE as an unsigned number (0xFF). We
can specify this value as -1 instead and sans other gremlins in the code,
this should now work as intended.