Fix a bizarre integer overflow error in the enemy randomizer
The value that was originally used was flat out not representable, so
the actual behavior was implementation defined. Recent versions of MSVC and
GCC have the same practical behavior of masking off the high bits until
the value is representable, so I've just made that explicit. No idea
what was originally intended though.